requirements.txt
future
emoji
requests==2.3.0
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
googletrans
Pillow>=4.2.0
aiohttp>=2.2.5
pynewtonmath
psutil
bleach
markdown2
speedtest-cli
wikipedia
pylast
jikanpy
coffeehouse
tswift
